Code clean up: move `BundleCheckpointURLProvider` to its own file
